+++ client/assets/scripts/editQs.js
