+++ client/assets/scripts/index.js
@@ -1,17 +1,17 @@
 const { TestEnvironment } = require("jest-environment-jsdom");
 
 function gameOne() {
-    window.location.href="/homePage/test-page.html";
+    window.location.href="/test-page.html";
 }
 
 function userRecord() {
     document.querySelector("#user-record").addEventListener("click", userRecord);
-    window.location.href="/homePage/user-record.html";
+    window.location.href="/user-record.html";
 }
 
 function leaderBoard() {
     document.querySelector("#leader-board").addEventListener("click", leaderBoard);
-    window.location.href="/homePage/leaderboard.html";
+    window.location.href="/leaderboard.html";
 }
 
 function heading() {
