+++ client/assets/scripts/results.js
@@ -3,15 +3,15 @@ console.log(queryString);
 const urlParams = new URLSearchParams(queryString);
 document.querySelector("#retry").addEventListener("click", qPage);
 function qPage() {
-    window.location.href="/questionPage/questionPage.html";
+    window.location.href="/questionPage.html";
 }
 document.querySelector("#home").addEventListener("click", hPage);
 function hPage() {
-    window.location.href="/homePage/index.html";
+    window.location.href="/index.html";
 }
 document.querySelector("#leader").addEventListener("click", lPage);
 function lPage() {
-    window.location.href="/questionPage/questionPage.html";
+    window.location.href="/questionPage.html";
 }
 function displayScore () {
         const card = document.createElement("div")
