+++ client/assets/scripts/settings.js
@@ -15,14 +15,14 @@ function handleSubmit(e) {
     const category = document.getElementById("category").value;
     const questions = document.getElementById("questions").value;
     
-    const newUrl = `/questionPage/questionPage.html?name=${name}&category=${category}&questions=${questions}`;
+    const newUrl = `/questionPage.html?name=${name}&category=${category}&questions=${questions}`;
   
     window.location.href = newUrl;
     console.log(newUrl)
   }
 
   function navigateToAddQuestions() {
-    window.location.href = "/editPage/editQs.html";
+    window.location.href = "/editQuestions.html";
 }
 
   module.exports = {
