+++ client/editQuestions.html
@@ -5,8 +5,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Edit Questions</title>
-    <script src="/editPage/editQs.js" defer></script>
-    <link rel="stylesheet" href="/editPage/editQs.css">
+    <script src="/assets/scripts/editQs.js" defer></script>
+    <link rel="stylesheet" href="/assets/style/editQs.css">
 </head>
 <body>
     <form>
