--- client/homePage/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "homepage",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "jest --environment=jsdom",
-    "gameTest": "jest --verbose index.test.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "jest-environment-jsdom": "^29.3.1"
-  }
-}
