--- client/homePage/user-record.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/homePage/index.css">
-    <title>User Records</title>
-</head>
-<body>
-   <h1>Your Records</h1> 
-   <table>
-    <tr>
-      <th>Rank</th>
-      <th>Game</th>
-      <th>Score</th>
-    </tr>
-    <tr>
-      <td>200</td>
-      <td>Game 1</td>
-      <td>940</td>
-    </tr>
-    <tr>
-      <td>348</td>
-      <td>Game 2</td>
-      <td>450</td>
-    </tr>
-    <tr>
-      <td>179</td>
-      <td>Game 3</td>
-      <td>900</td>
-    </tr>
-  </table>
-</body>
-</html>
