+++ client/index.html
@@ -4,9 +4,9 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/homePage/index.css">
-    <script src="/homePage/test-page.html"></script>
-    <script src="/homePage/index.js"></script>
+    <link rel="stylesheet" href="/assets/style/index.css">
+    <script src="/assets/scripts/test-page.html"></script>
+    <script src="/assets/scripts/index.js"></script>
     <title>Home</title>
 </head>
 <body>
