+++ client/leaderboard.html
@@ -1,0 +1,36 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="/assets/scripts/index.css">
+    <script src="/assets/scripts/index.js"></script>
+    <title>Leaderboard</title>
+</head>
+<body>
+   <h1>LeaderBoard</h1> 
+   <table>
+    <tr>
+      <th>Rank</th>
+      <th>Username</th>
+      <th>Score</th>
+    </tr>
+    <tr>
+      <td>1</td>
+      <td>player1</td>
+      <td>1000</td>
+    </tr>
+    <tr>
+      <td>2</td>
+      <td>player2</td>
+      <td>900</td>
+    </tr>
+    <tr>
+      <td>3</td>
+      <td>player3</td>
+      <td>800</td>
+    </tr>
+  </table>
+</body>
+</html>
