+++ client/questionPage.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/questionPage/questionPage.css">
-    <script src="/questionPage/question.js" defer></script>
+    <link rel="stylesheet" href="/assets/style/questionPage.css">
+    <script src="/assets/scripts/question.js" defer></script>
     <title>Quiz</title>
 </head>
 <body>
