+++ client/results.html
@@ -5,8 +5,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <link rel="stylesheet" href="/resultPage/results.css">
-    <script src="/resultPage/results.js" defer></script>
+    <link rel="stylesheet" href="/assets/style/results.css">
+    <script src="/assets/scripts/results.js" defer></script>
 </head>
 <body>
 <h1>Results</h1>
