+++ client/settings.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/settingsPage/settings.css">
-    <script src="/settingsPage/settings.js" defer></script>
+    <link rel="stylesheet" href="/assets/style/settings.css">
+    <script src="/assets/scripts/settings.js" defer></script>
     <title>Settings</title>
 </head>
 <body>
