+++ client/user-record.html
@@ -4,32 +4,31 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/homePage/index.css">
-    <script src="/homePage/index.js"></script>
-    <title>Leaderboard</title>
+    <link rel="stylesheet" href="/assets/style/index.css">
+    <title>User Records</title>
 </head>
 <body>
-   <h1>LeaderBoard</h1> 
+   <h1>Your Records</h1> 
    <table>
     <tr>
       <th>Rank</th>
-      <th>Username</th>
+      <th>Game</th>
       <th>Score</th>
     </tr>
     <tr>
-      <td>1</td>
-      <td>player1</td>
-      <td>1000</td>
+      <td>200</td>
+      <td>Game 1</td>
+      <td>940</td>
     </tr>
     <tr>
-      <td>2</td>
-      <td>player2</td>
+      <td>348</td>
+      <td>Game 2</td>
+      <td>450</td>
+    </tr>
+    <tr>
+      <td>179</td>
+      <td>Game 3</td>
       <td>900</td>
-    </tr>
-    <tr>
-      <td>3</td>
-      <td>player3</td>
-      <td>800</td>
     </tr>
   </table>
 </body>
